Refactor component imports to use correct context path
