+++ client/components/Nav/Nav.jsx
@@ -3,7 +3,7 @@ import NavLink from "./NavLink";
 import Button from "../Button";
 
 import Link from "next/link";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../../context/AuthContext";
 
 const navData = {
   links: [
