+++ client/pages/_app.js
@@ -1,4 +1,4 @@
-import { AuthProvider } from "@/components/context/AuthContext";
+import { AuthProvider } from "@/context/AuthContext";
 import "@/styles/globals.css";
 
 export default function App({ Component, pageProps }) {
